+++ test/dummy/db/migrate/20260910020000_add_icon_to_recording_studio_support_pages.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class AddIconToRecordingStudioSupportPages < ActiveRecord::Migration[8.1]
+  def change
+    add_column :recording_studio_support_pages, :icon, :string
+  end
+end